+++ src/bases/http-client/RequestCancelledError.test.ts
